requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
pyyaml
python-dateutil
pytz
beautifulsoup4
tqdm
colorama
